Add files via upload
This is prepared for the working paper "Peer Effects on Altruistic Behaviors: Examination of Spatial, Social, and Economic Networks with the Data of the Rice Farmers in Bohol, the Philippines". It is one sample of the overall 19 do-files. The other do files are similar but based on different matrixes. There are 19 different spatial weight matrices in total measured by different relationships among the villagers such as meet daily, go to weddings, and so on.